Fix HWND result from IShellWindows::FindWindowSW
- Initialize HWND hwnd before IShellWindows::FindWindowSW writes to it.

In GetShellViewForDesktop the HWND hwnd receives the desktop Window from
IShellWindows::FindWindowSW which only writes 32 bits into hwnd. Prior
to this change in 64-bit builds (which have 64-bit HWND type) the upper
32 bits of hwnd remained uninitialized.

This change was made for best practice. Windows API ignores the upper 32
bits of the HWND type for compatibility reasons and currently HWND hwnd
isn't actually used after it's written to.